+++ backend/src/baserow/config/settings/test.py
@@ -1,3 +1,5 @@
+from copy import deepcopy
+
 from .base import *  # noqa: F403, F401
 
 
@@ -9,6 +11,9 @@ CELERY_TASK_ALWAYS_EAGER = True
 CELERY_TASK_EAGER_PROPAGATES = True
 
 CHANNEL_LAYERS = {"default": {"BACKEND": "channels.layers.InMemoryChannelLayer"}}
+
+# Open a second database connection that can be used to test transactions.
+DATABASES["default-copy"] = deepcopy(DATABASES["default"])  # noqa: F405
 
 USER_FILES_DIRECTORY = "user_files"
 USER_THUMBNAILS_DIRECTORY = "thumbnails"