+++ backend/src/baserow/contrib/database/models.py
@@ -31,9 +31,10 @@ from .webhooks.models import (
     TableWebhookCall,
     TableWebhookHeader,
 )
-from baserow.contrib.database.fields.dependencies.models import (
-    FieldDependency,
-)
+from .airtable.models import AirtableImportJob
+
+from baserow.contrib.database.fields.dependencies.models import FieldDependency
+
 
 __all__ = [
     "Database",
@@ -63,6 +64,7 @@ __all__ = [
     "TableWebhookEvent",
     "TableWebhookHeader",
     "TableWebhookCall",
+    "AirtableImportJob",
     "FieldDependency",
 ]