+++ web-frontend/modules/database/components/view/grid/GridView.vue
@@ -1,5 +1,9 @@
 <template>
-  <div v-scroll="scroll" class="grid-view">
+  <div
+    v-scroll="scroll"
+    class="grid-view"
+    :class="{ 'grid-view--disable-selection': isMultiSelectActive }"
+  >
     <Scrollbars
       ref="scrollbars"
       horizontal="getHorizontalScrollbarElement"
@@ -24,6 +28,9 @@
       @row-hover="setRowHover($event.row, $event.value)"
       @row-context="showRowContext($event.event, $event.row)"
       @row-dragging="rowDragStart"
+      @cell-mousedown-left="multiSelectStart"
+      @cell-mouseover="multiSelectHold"
+      @cell-mouseup-left="multiSelectStop"
       @add-row="addRow()"
       @update="updateValue"
       @edit="editValue"
@@ -31,6 +38,7 @@
       @unselected="unselectedCell($event)"
       @select-next="selectNextCell($event)"
       @edit-modal="$refs.rowEditModal.show($event.id)"
+      @scroll="scroll($event.pixelY, 0)"
     >
       <template #foot>
         <div class="grid-view__foot-info">
@@ -70,6 +78,9 @@
       @add-row="addRow()"
       @update="updateValue"
       @edit="editValue"
+      @cell-mousedown-left="multiSelectStart"
+      @cell-mouseover="multiSelectHold"
+      @cell-mouseup-left="multiSelectStop"
       @selected="selectedCell($event)"
       @unselected="unselectedCell($event)"
       @select-next="selectNextCell($event)"
@@ -103,7 +114,15 @@
       @scroll="scroll($event.pixelY, $event.pixelX)"
     ></GridViewRowDragging>
     <Context ref="rowContext">
-      <ul class="context__menu">
+      <ul v-show="isMultiSelectActive" class="context__menu">
+        <li>
+          <a @click=";[exportMultiSelect(), $refs.rowContext.hide()]">
+            <i class="context__menu-icon fas fa-fw fa-copy"></i>
+            {{ $t('action.copy') }}
+          </a>
+        </li>
+      </ul>
+      <ul v-show="!isMultiSelectActive" class="context__menu">
         <li v-if="!readOnly">
           <a @click=";[addRow(selectedRow), $refs.rowContext.hide()]">
             <i class="context__menu-icon fas fa-fw fa-arrow-up"></i>
@@ -163,6 +182,7 @@ import GridViewRowDragging from '@baserow/modules/database/components/view/grid/
 import RowEditModal from '@baserow/modules/database/components/row/RowEditModal'
 import gridViewHelpers from '@baserow/modules/database/mixins/gridViewHelpers'
 import { maxPossibleOrderValue } from '@baserow/modules/database/viewTypes'
+import { isElement } from '@baserow/modules/core/utils/dom'
 import viewHelpers from '@baserow/modules/database/mixins/viewHelpers'
 
 export default {
@@ -277,6 +297,8 @@ export default {
       ...mapGetters({
         allRows: this.$options.propsData.storePrefix + 'view/grid/getAllRows',
         count: this.$options.propsData.storePrefix + 'view/grid/getCount',
+        isMultiSelectActive:
+          this.$options.propsData.storePrefix + 'view/grid/isMultiSelectActive',
       }),
     }
   },
@@ -297,6 +319,14 @@ export default {
     }
     this.$el.resizeEvent()
     window.addEventListener('resize', this.$el.resizeEvent)
+    window.addEventListener('keydown', this.arrowEvent)
+    window.addEventListener('copy', this.exportMultiSelect)
+    window.addEventListener('click', this.cancelMultiSelect)
+    window.addEventListener('mouseup', this.multiSelectStop)
+    this.$refs.left.$el.addEventListener(
+      'scroll',
+      this.$el.horizontalScrollEvent
+    )
     this.$store.dispatch(
       this.storePrefix + 'view/grid/fetchAllFieldAggregationData',
       { view: this.view }
@@ -304,7 +334,16 @@ export default {
   },
   beforeDestroy() {
     window.removeEventListener('resize', this.$el.resizeEvent)
+    window.removeEventListener('keydown', this.arrowEvent)
+    window.removeEventListener('copy', this.exportMultiSelect)
+    window.removeEventListener('click', this.cancelMultiSelect)
+    window.removeEventListener('mouseup', this.multiSelectStop)
     this.$bus.$off('field-deleted', this.fieldDeleted)
+    this.$store.dispatch(
+      this.storePrefix + 'view/grid/setMultiSelectActive',
+      false
+    )
+    this.$store.dispatch(this.storePrefix + 'view/grid/clearMultiSelect')
   },
   methods: {
     /**
@@ -704,6 +743,12 @@ export default {
         return
       }
 
+      this.$store.dispatch(
+        this.storePrefix + 'view/grid/setMultiSelectActive',
+        false
+      )
+      this.$store.dispatch(this.storePrefix + 'view/grid/clearMultiSelect')
+
       this.$store.dispatch(this.storePrefix + 'view/grid/setSelectedCell', {
         rowId: nextRowId,
         fieldId: nextFieldId,
@@ -722,6 +767,95 @@ export default {
       this.$nextTick(() => {
         this.fieldsUpdated()
       })
+    },
+    /*
+      Called when mouse is clicked and held on a GridViewCell component.
+      Starts multi-select by setting the head and tail index to the currently
+      selected cell.
+    */
+    multiSelectStart({ event, row, field }) {
+      this.$store.dispatch(this.storePrefix + 'view/grid/multiSelectStart', {
+        rowId: row.id,
+        fieldIndex: this.visibleFields.findIndex((f) => f.id === field.id) + 1,
+      })
+    },
+    /*
+      Called when mouse hovers over a GridViewCell component.
+      Updates the current multi-select grid by updating the tail index
+      with the last cell hovered over.
+    */
+    multiSelectHold({ event, row, field }) {
+      this.$store.dispatch(this.storePrefix + 'view/grid/multiSelectHold', {
+        rowId: row.id,
+        fieldIndex: this.visibleFields.findIndex((f) => f.id === field.id) + 1,
+      })
+    },
+    /*
+      Called when the mouse is unpressed over a GridViewCell component.
+      Stop multi-select.
+    */
+    multiSelectStop({ event, row, field }) {
+      this.$store.dispatch(
+        this.storePrefix + 'view/grid/setMultiSelectHolding',
+        false
+      )
+    },
+    /*
+      Cancels multi-select if it's currently active.
+      This function checks if a mouse click event is triggered
+      outside of GridViewRows. This is done by ensuring that the
+      target element's class is either 'grid-view' or 'grid-view__rows'.
+    */
+    cancelMultiSelect(event) {
+      if (
+        this.$store.getters[
+          this.storePrefix + 'view/grid/isMultiSelectActive'
+        ] &&
+        (!isElement(this.$el, event.target) ||
+          !['grid-view__rows', 'grid-view'].includes(event.target.classList[0]))
+      ) {
+        this.$store.dispatch(
+          this.storePrefix + 'view/grid/setMultiSelectActive',
+          false
+        )
+        this.$store.dispatch(this.storePrefix + 'view/grid/clearMultiSelect')
+      }
+    },
+    arrowEvent(event) {
+      // Check if arrow key was pressed.
+      if (
+        ['ArrowRight', 'ArrowLeft', 'ArrowUp', 'ArrowDown'].includes(event.key)
+      ) {
+        // Cancels multi-select if it's currently active.
+        if (
+          this.$store.getters[
+            this.storePrefix + 'view/grid/isMultiSelectActive'
+          ]
+        ) {
+          this.$store.dispatch(
+            this.storePrefix + 'view/grid/setMultiSelectActive',
+            false
+          )
+          this.$store.dispatch(this.storePrefix + 'view/grid/clearMultiSelect')
+        }
+      }
+    },
+    // Prepare and copy the multi-select cells into the clipboard, formatted as TSV
+    async exportMultiSelect(event) {
+      try {
+        this.$store.dispatch('notification/setCopying', true)
+        const output = await this.$store.dispatch(
+          this.storePrefix + 'view/grid/exportMultiSelect',
+          this.leftFields.concat(this.visibleFields)
+        )
+        if (output !== undefined) {
+          navigator.clipboard.writeText(output)
+        }
+      } catch (error) {
+        notifyIf(error, 'view')
+      } finally {
+        this.$store.dispatch('notification/setCopying', false)
+      }
     },
   },
 }